+++ routes/index.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { login, createUser } = require('../controllers/users');
+const { login, createUser } = require('../controllers/user');
 const auth = require('../middlewares/auth');
 const userRouter = require('./users');
 const articleRouter = require('./articles');
